log discovery every 30 minutes in discovery processor
