+++ internal/plugincommon/discovery/processor.go
@@ -27,6 +27,7 @@ var _ plugincommon.PluginProcessor[dt.Query, dt.Observation, dt.Outcome] = &Cont
 
 // syncTimeout is the timeout for the ccip reader sync operation.
 const syncTimeout = 5 * time.Second
+const logFrequency = 30 * time.Minute // how often to log discovery progress
 
 // ContractDiscoveryProcessor is a plugin processor for discovering contracts.
 type ContractDiscoveryProcessor struct {
@@ -38,6 +39,7 @@ type ContractDiscoveryProcessor struct {
 	oracleIDToP2PID map[commontypes.OracleID]ragep2ptypes.PeerID
 	syncer          *readerSyncer
 	peerID          ragep2ptypes.PeerID
+	lastLogTime     atomic.Pointer[time.Time] // used to limit logging frequency
 }
 
 func NewContractDiscoveryProcessor(
@@ -294,11 +296,14 @@ func (cdp *ContractDiscoveryProcessor) Outcome(
 			agg.onrampAddrs,
 			destThresh,
 		)
-		lggr.Debugw("Determined consensus onramps",
-			"onrampConsensus", onrampConsensus,
-			"onrampAddrs", agg.onrampAddrs,
-			"fChainThresh", fChainThresh,
-		)
+		logutil.LogWhenExceedFrequency(
+			&cdp.lastLogTime, logFrequency, func() {
+				lggr.Debugw("Determined consensus onramps",
+					"onrampConsensus", onrampConsensus,
+					"onrampAddrs", agg.onrampAddrs,
+					"fChainThresh", fChainThresh,
+				)
+			})
 		if len(onrampConsensus) == 0 {
 			lggr.Warnw("No consensus on onramps, onrampConsensus map is empty")
 		}
@@ -328,11 +333,14 @@ func (cdp *ContractDiscoveryProcessor) Outcome(
 		agg.rmnRemoteAddrs,
 		fChainThresh,
 	)
-	lggr.Debugw("Determined consensus RMNRemote",
-		"rmnRemoteConsensus", rmnRemoteConsensus,
-		"rmnRemoteAddrs", agg.rmnRemoteAddrs,
-		"fChainThresh", fChainThresh,
-	)
+	logutil.LogWhenExceedFrequency(
+		&cdp.lastLogTime, logFrequency, func() {
+			lggr.Debugw("Determined consensus RMNRemote",
+				"rmnRemoteConsensus", rmnRemoteConsensus,
+				"rmnRemoteAddrs", agg.rmnRemoteAddrs,
+				"fChainThresh", fChainThresh,
+			)
+		})
 	if len(rmnRemoteConsensus) == 0 {
 		lggr.Warnw("No consensus on RMNRemote, rmnRemoteConsensus map is empty")
 	}
@@ -344,11 +352,14 @@ func (cdp *ContractDiscoveryProcessor) Outcome(
 		agg.feeQuoterAddrs,
 		fChainThresh,
 	)
-	lggr.Debugw("Determined consensus fee quoter",
-		"feeQuoterConsensus", feeQuoterConsensus,
-		"feeQuoterAddrs", agg.feeQuoterAddrs,
-		"fChainThresh", fChainThresh,
-	)
+	logutil.LogWhenExceedFrequency(
+		&cdp.lastLogTime, logFrequency, func() {
+			lggr.Debugw("Determined consensus fee quoter",
+				"feeQuoterConsensus", feeQuoterConsensus,
+				"feeQuoterAddrs", agg.feeQuoterAddrs,
+				"fChainThresh", fChainThresh,
+			)
+		})
 	if len(feeQuoterConsensus) == 0 {
 		lggr.Warnw("No consensus on fee quoters, feeQuoterConsensus map is empty")
 	}
@@ -361,11 +372,14 @@ func (cdp *ContractDiscoveryProcessor) Outcome(
 		agg.routerAddrs,
 		fChainThresh,
 	)
-	lggr.Debugw("Determined consensus router",
-		"routerConsensus", routerConsensus,
-		"routerAddrs", agg.routerAddrs,
-		"fChainThresh", fChainThresh,
-	)
+	logutil.LogWhenExceedFrequency(
+		&cdp.lastLogTime, logFrequency, func() {
+			lggr.Debugw("Determined consensus router",
+				"routerConsensus", routerConsensus,
+				"routerAddrs", agg.routerAddrs,
+				"fChainThresh", fChainThresh,
+			)
+		})
 	if len(routerConsensus) == 0 {
 		lggr.Warnw("No consensus on router, routerConsensus map is empty")
 	}
